I added a bunch of bogus bond and angle terms so that the simulator could handle the fine motion controller without hiccuping. I'd chosen the bond and angle coefficients pretty randomly (just cutting and pasting values from unrelated previous entries). Here I've tried to make the coefficients small so that these bogus new entries have minimal effects.
Probably the right thing to do is, for each bond/angle situation we hit
for which we don't have an existing entry, we use a default (very weak)
entry and (on the first instance only, hopefully) we issue a warning.
That would mean keeping track of which novel bond/angle situations we're
presented with, which is (a) memory and (b) programming effort. I dunno
if it's worth it.