Rework the simulator test framework to use Python's unit test facility. This means that simulator tests can later be integrated easily into a larger group of whole-system tests, when we're ready for something like that, and can be a little easier to understand in the meanwhile.
The reason for checking in subprocess.py (which comes straight out
of Python 2.4's lib) is that I'm running 2.3 here, as I imagine
others may be doing, and I don't want to force people to switch
when subprocess.py will work just fine on Python 2.3.

Alternatively, people could copy it into their 2.3 library directory,
if it seems undesirable to clutter the sim/src directory with it.