This checkin makes the simulator reentrant, or at least is the first step toward doing so. A context struct is declared in globals.h. Currently it is only the new residence of the ToMinimize flag, which was previously global, but other variables may be migrated there later. The motivation is to make the simulator a more natural Python extension.
Since this is a big checkin, even though I'm pretty confident that it
shouldn't break anything, I did a few minimizations and dynamics runs with
the CAD program, and verified that the test coverage in tests.py hadn't
changed. If we do need to revert it, I set up a non-branch CVS tag at the
top of the sim tree called "wware_pre_reentrancy_20051219".