Emacs Python mode gets confused if you use a single quote inside a multi-line docstring delineated by single quotes. Change delineators to double quotes; less ugly than writing "the new node\'s name...".
Scrap _just_updated, which never worked and has become obsolete. Scrap
_create_new_name because nobody needs to overload it.

Change the build_struct method to take "name" and "position"
parameters in addition to the tuple of standard parameters. This
cleans up some code, and allows us to position the generated object,
so that we can generate objects that don't overlap becausse they don't
need to be centered at (0,0,0).