Only use the thick lines when dragging the jig. Put back the arrowheads because they look good with thin lines. Fix a bug in the positioning of arrowheads on angle and dihedral measurements.
There is an artifact that appears on thick lines for angle and dihedral
measurements. A small wedge seems to be chopped out of the thick line.
This happens at 45 degrees from the corner of the angle, and it happens
because the little lines making up the arc are switching between vertical
and horizontal. I don't think there is anything we can do about this.