In fileExport, reinstate Mark's code for exporting PDBs rather than using Open Babel. I was getting a little ahead of myself, there are still cases there where Open Babel chokes.
In fileExprot, replace Popen3 (not available on Windows) with
QProcess, which provides both exit status and standard error.

When exporting with a file extension of "MDL", do not use the
chemistry format that Open Babel knows about. Use the Animation
Master format that John Burch used to make the nanofactory
animation.