The Numeric array formatting (array_repr) for this one little assert in Chunk.draw was eating a sixth of a second out of the one second rotate redraw time for the FourHoleTile with only the staple strands drawn.  It was string-formatting 4 arrays (2 vectors and 2 quaterions) with back-ticks and doing a string comparison.
I just stuck an "if !=" test in front of the assert so it wouldn't do the formatting unless needed.

Bruce, I did enough testing on the relevant comparison case (tuple of a vector and a quaternion) to convince myself that a "!=" test would work and removed this comment:
    # (we use `x` == `y` since x == y doesn't work well for these data types)
Tell me if you know of a case this doesn' handle.