Almost completely remove the "DNA strand vanishing" bug, which only appears on large DNAO tiles like the FourHoleTile and smiley-face, at the very beginning of inserting the first additional DNA strand.  Instead of most of 19 strands vanishing, there's now only one chunk of one strand gone on the FourHoleTile.
The only significant change is in CSDL.reset(), which puts a CSDL back in initialized state now rather than activated state, clearing it first if necessary.  CSDL.activate only allocates a single display list id, which was then never used, so it's not clear why this makes any difference!

CSDL.deallocate_displists is also careful now not to double-delete self.dl when it is a duplicate of self.color_dl or self.selected_dl, but that doesn't affect the strands vanishing.

Also numerous ignorable clarifications and some reformatting of comments.