Fix the highlightTimer, which yanks away window focus when it fires on a buried window.  Treat focusIn/focusOut events the same as enter/leave events.
On the Mac at least, Cmd-Tabbing to another app window that pops up on
top of our GLPane doesn't deliver a leave event, but does deliver a
focusOut.  Unless we handle it as a leave, the timer is left active, and
a highlight draw can occur.  This steals the focus from the upper
window, popping NE1 on top of it, which is very annoying to the user.

Here's a way to reproduce it:

  1. Open NE1, and put it over another app window, burying it.  (It
     doesn't seem to matter which app is used.)
  2. Put anything that will hover-highlight into the GLPane.  Make it
     nice and big, since you'll be moving the mouse 'blind'.
  3. Position the mouse: hover-highlight the target structure bright
     yellow (or put the mouse right next to something that will
     hover-hightlight when you move the mouse over it.)  Don't move the
     mouse yet.
  4. Cmd-Tab to the buried app, taking the focus away from NE1.
  5. Nudge the mouse just a little bit and wait a second.
     Hover-highlighting is triggered in the buried NE1 window, popping
     it to the top and yanking the focus away from the other app.

You can watch what's happening by setting DEBUG_BAREMOTION in
cad/src/utilities/GlobalPreferences.py .