Improve the experimental CSDL.draw() method.
Change the default patterning option to True, since this is more often going to be
used for things that are highlighted like chunks than for things that are highlighted
like atoms.

Add an option to over-ride the hover-highlight color, because it's easy to do so.

Improve the comments, in an attempt to clarify the interaction of the many cases this
handles.  (Highlighting/Selection/Ordinary drawing, with patterned or solid
highlighting and/or selection overlays.)

Correct a logic error for objects that are both selected and highlighted.

Still untested; no guarantee that this code does anything at all beyond parsing correctly.