Fix patterned drawing traceback.
Changed the comparison for "pattern" from "is False" to "== False" to avoid a potential bug.  That introduced another bug in this case, since the value is a numpy array when it is set.  The numpy array == operator doesn't accept an array of more than one element as a boolean value.

Change the unset value to None, and the comparison to "is None".