graphics/drawing/gl_shaders.py is loaded by setup_draw.py, but nothing new is done unless the OpenGL extension GL_ARB_shader_objects exists.  If it is, drawing_globals.sphereShader is created.
exprs/test_drawing.py is not even loaded unless _testing_ is enabled at the beginning of graphics/widgets/glPane.py .

When it is, GLPane.paintGL() calls the test_drawing() function instead of the usual body of paintGL().  It draws a square array of spheres in a variety of different ways, and prints a frames-per-second counter.

The testCase and nSpheres variables at the beginning of test_drawing.py control which test is run, and how large an array of spheres is drawn.  (No debug prefs yet...)

See the test cases in test_drawing() for descriptions and timings on a MacBook Pro.