Support for transforms in drawing-primitive shader batches
An optional 'transform_id' indexes transform matrices applied by drawing-primitive vertex shaders.  This will be used by TransformControl objects that group CSDLs for dragging, etc.

Transform_ids are optionally set per drawing primitive in batches, and stored as per-vertex attributes in VBOs in graphics card RAM.

The transform matrices are stored in graphics card RAM in a (potentially large) block of texture memory, or in a smaller fixed-sized block of uniform (constant) memory.

The constant 'texture_xforms' in gl_shaders.py controls the storage mode.  Texture memory transforms appear to be working now and scale well, so that's the default.  The constant array is much simpler, so it's left in for now if needed in further development, but automatically ifdef'ed out when texture_xforms is set.