Minor changes to batch-shader drawing primitives
It turns out I was wrong; transform matrices in texture memory still return zeros.  I faked my self out with a side-effect of some shader debug code that was not within the #if.  That is fixed now, and the default is now switched back to the smaller constant array.

Fix a bug: transform the center_pt vertex attributes as well as the vertex coordinates.

In debugging code, add a new pattern that labels all matrix elements, and format matrices read back from texture memory better.