Clean up an import cycle kluge in drawing_globals and glprefs.
drawing_globals.glprefs was being set up as a side effect of an extraneous import of glprefs in GLPane_rendering_methods.  This was ugly and confusing.

Now it's all handled between drawing_globals and glprefs, although there is still a run-time import cycle requiring a couple of imports within functions.  Unideal but ok for now, in Bruce's opinion.