A future bug catcher and minor documentation changes.
-- graphics/drawing/ColorSorter.py
Complain of a bug if the ColorSorter attempts to push a null name on the glPushName stack.
This code has silently ignored zero or None glnames since it was written in rev 2943, 16 Feb 2006.
It turned out not to be the cause of bug 2945, but should be protected against.

-- graphics/widgets/GLPane_rendering_methods.py
Bruce pointed out that the 'glselect_glname_color' drawing_phase also restricts the viewport (to one pixel.)

-- graphics/drawing/gl_shaders.py
Bruce also pointed out that configShader should contain discussion of two related issues:
. Consider caching the glpane pointer passed to it as an argument.
. Multiple OpenGL contexts that don't share handles, e.g. display lists, in the future.