Use CSDL.draw() in more places.
Calls on CSDL.draw_dl() or glCallList with Display List handles stored in CSDLs are being converted to CSDL.draw().  This fixed several bugs where spheres were drawn as triangles, or shader spheres disappeared from the screen.

Also fix the problem where shader spheres were missing on the first draw after (re-)filling a CSDL with the ColorSorter.  This was simply due to the design that the ColorSorter is supposed to both build and draw the contents of a CSDL as it's filled.

The ColorSorter used glNewList(GL_COMPILE_AND_EXECUTE) to build single-level Display Lists: they were getting drawn as they were built, but the shader spheres aren't in DLs and so weren't.  Changed to GL_COMPILE mode, and added a call on CSDL.draw() at the end of ColorSorter.finish().

Multi-level DLs are built in GL_COMPILE mode, but were drawn with CSDL.draw_dl(), which also ignored shader spheres.  Removed the call on draw_dl() from this case, since the call on CSDL.draw() at the end does the drawing properly.