Fix exception when starting sphere shaders on Windows.
In GLSphereShaderObject.createShader(),
        glShaderSourceARB(shader, shaderSrc)
throws:
    exceptions.TypeError: this function takes 4 arguments (2 given)

The C calling sequence does indeed take 4 arguments, but the last three are all about an array of strings: count, string pointers, and string lengths.  In the PyOpenGL wrapper, a single argument containing a list of strings produces the three C args, so there are only two arguments in the Python calling sequence.

It worked fine on MacOS SOs, but the wrappers weren't getting defined on Windows, due differences in the way PyOpenGL loads DLLs.  This left the "raw" Python function definitions exposed, expecting the C arguments.  (It's confusing that all of the different functions have the same names.)

This was broken in PyOpenGL 3.0.0a6, but fixed in 3.0.0b3 .  The shader_objects_patch.py and vertex_shader_patch.py files added here are from the 3.0.0b3 sources, slightly modified to fit into 3.0.0a6 .

We can't yet update to PyOpenGL 3.0.0b3 due to other incompatibilities that need to be fixed, but if we had, this wouldn't have been a problem.