Fix a cylinder vertex shader perspective billboard bug.
The near end of a cylinder billboard was a little bit narrow when perspective was on, so slivers of the edges of the cylinder barrel were sliced away by the billboard edges.

The fix is a proper projection of the vertices shared by the barrel and visible endcap surfaces, along lines coming from the viewpoint, to the far side of the endcap.  This makes the billboard wider by the proper amount for any viewpoint.