Fix cylinder vertex shader drawing bugs.
(1.) In half of the viewing orientations, with both perspective and orthographic viewing, the cylinder barrel became flattened toward the axis and appeared truncated, but still rounded, as the viewing direction approached the axis direction.

This was difficult to diagnose because the nearer endcap blocked view of the flattened barrel behind it.  Disabling endcap drawing made the cause visible.

This was a simple case of forgetting that the axis direction can be opposite to the view direction.  I used the length of a vector (proj_passing_dist), losing that direction sign bit.  The fix is obvious.

(2.) In perspective viewing, when the back side of the first endcap is visible (the one at the local coordinate system end of the cylinder), half of the endcap pixels were shown at each end of the cylinder in a moire pattern, even when both endcaps should be hidden.

This turned out to be two bugs related to the same cause.  There are two "varying" integer codes passed from the vertex shader to the pixel shader.  They are var_visible_endcap (0 or 1) and var_visibility (0 to 3).

Varyings are always floats, and are interpolated between vertices.  Even though the same integer value goes into these variables at each vertex, the interpolated result is not *exactly* the same integer.  Sometimes it was just above the integer value, and sometimes just below, wreaking havoc when the int() conversion truncated the value.

 . The fix is to simply add 0.5 to round the float value when converting back to an int.

 . The int name of var_visible_endcap is visible_endcap.  There wasn't one for var_visible, so change float var_visible references to int visibility_type.